+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fix download/share button showing in overdue search screen outside India
 - Add support for Android 13
   - Bump compile and target SDK to 32
+- Bump fragment to v1.5.2
 
 ## 2022-08-08-8359